Refactored CustomRequestTest class
